+++ sdk/messaging/azeventhubs/processor.go
@@ -10,8 +10,13 @@ import (
 	"sync"
 	"time"
 
+	"github.com/Azure/azure-sdk-for-go/sdk/azcore/to"
 	azlog "github.com/Azure/azure-sdk-for-go/sdk/internal/log"
 )
+
+// processorOwnerLevel is the owner level we assign to every ProcessorPartitionClient
+// created by this Processor.
+var processorOwnerLevel = to.Ptr[int64](0)
 
 // ProcessorStrategy specifies the load balancing strategy used by the Processor.
 type ProcessorStrategy string
@@ -49,13 +54,6 @@ type ProcessorOptions struct {
 	// The default position is Latest.
 	StartPositions StartPositions
 
-	// OwnerLevel is the priority for partition clients created by this Processor, also known as
-	// the 'epoch' level.
-	// When used, a partition client with a higher OwnerLevel will take ownership of a partition
-	// from partition clients with a lower OwnerLevel.
-	// Default is 0.
-	OwnerLevel int64
-
 	// Prefetch represents the size of the internal prefetch buffer for each ProcessorPartitionClient
 	// created by this Processor. When set, this client will attempt to always maintain
 	// an internal cache of events of this size, asynchronously, increasing the odds that
@@ -91,7 +89,6 @@ type Processor struct {
 	ownershipUpdateInterval time.Duration
 	defaultStartPositions   StartPositions
 	checkpointStore         CheckpointStore
-	ownerLevel              int64
 	prefetch                int32
 
 	// consumerClient is actually a *azeventhubs.ConsumerClient
@@ -158,7 +155,6 @@ func newProcessorImpl(consumerClient consumerClientForProcessor, checkpointStore
 	}
 
 	return &Processor{
-		ownerLevel:              options.OwnerLevel,
 		ownershipUpdateInterval: updateInterval,
 		consumerClient:          consumerClient,
 		checkpointStore:         checkpointStore,
@@ -338,7 +334,7 @@ func (p *Processor) addPartitionClient(ctx context.Context, ownership Ownership,
 
 	partClient, err := p.consumerClient.NewPartitionClient(ownership.PartitionID, &PartitionClientOptions{
 		StartPosition: sp,
-		OwnerLevel:    &p.ownerLevel,
+		OwnerLevel:    processorOwnerLevel,
 		Prefetch:      p.prefetch,
 	})